% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Policy.R
\name{create_policy_RM}
\alias{create_policy_RM}
\title{Reverse Mortgage - Policy Object Generator}
\usage{
create_policy_RM(value, LVR, trans_cost, margin)
}
\arguments{
\item{value}{}

\item{LVR}{}

\item{trans_cost}{}

\item{margin}{}
}
\value{
Policy object
}
\description{
Generates policy object for Reverse Mortgage policies
}
\examples{
rm <- create_policy_RM(100000, 0.4, 0.01, 0.05)
}
