% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valuation.R
\name{get_path_prices}
\alias{get_path_prices}
\title{Valuation via cashflows}
\usage{
get_path_prices(cashflows)
}
\arguments{
\item{cashflows}{Matrix of simulated cashflow paths}

\item{ret_paths}{If ret_paths = TRUE, vector of prices for each path is returned}
}
\value{
Value of policy
}
\description{
Simulate cash flows using Monte-Carlo methods for various policies
}
\examples{
cf <- cashflow(policy = "VA", age = 65, sex = "M", n = 1000)
}
