% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valuation.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Valuation via cashflows}
\usage{
plot_distribution(prices)
}
\arguments{
\item{cashflows}{Matrix of simulated cashflow paths}

\item{convergence}{If convergence = TRUE, plots convergence of price}
}
\value{
Value of policy
}
\description{
Simulate cash flows using Monte-Carlo methods for various policies
}
\examples{
cf <- cashflow(policy = "VA", age = 65, sex = "M", n = 1000)
}
