% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cashflow.R
\name{cashflow}
\alias{cashflow}
\title{Policy Cashflow Simulator}
\usage{
cashflow(policy = "AP", age = 17, sex = "F", seed = 0, n = 1000)
}
\arguments{
\item{policy}{Policy type to simulate:
\code{policy = "AP"} (Account Based Pension),
\code{"RM"} (Reverse Mortgage),
\code{"LA"} (Life Annuity),
\code{"CA"} (Care Annuity),
\code{"PA"} (Pooled Annuity),
\code{"VA"} (Variable Annuity)}

\item{age}{Initial age of policyholder in years}

\item{sex}{sex of policyholder, \code{sex = "F"} (female), \code{"M"} (male)}

\item{seed}{Seed for random generator}

\item{n}{Number of paths to simulate (Monte-Carlo method)}
}
\value{
Matrix of cash flow vectors for each simulated path
}
\description{
Simulate cash flows using Monte-Carlo methods for various policies
}
\examples{
cf <- cashflow(policy = "VA", age = 65, sex = "M", n = 1000)
}
